Move IFolder out of GlobalFileSystem.cs
